+++ Data/Database/SQLite/DataModels/CREATETABLE/ProgramFinancingSchedule.sql
@@ -16,9 +16,9 @@ CREATE TABLE IF NOT EXISTS "ProgramFinancingSchedule"
 	"AccountSequence"	TEXT(80) NULL DEFAULT 'NS',
 	"AgencyName"	TEXT(80) NULL DEFAULT 'NS',
 	"FundName"	TEXT(80) NULL DEFAULT 'NS',
-	"OriginalAmount"	DECIMAL NULL DEFAULT 0,
-	"BudgetAmount"	DECIMAL NULL DEFAULT 0,
-	"AgencyAmount"	DECIMAL NULL DEFAULT 0,
-	"Amount"	DECIMAL NULL DEFAULT 0,
+	"OriginalAmount"	DOUBLE NULL DEFAULT 0.0,
+	"BudgetAmount"	DOUBLE NULL DEFAULT 0.0,
+	"AgencyAmount"	DOUBLE NULL DEFAULT 0.0,
+	"Amount"	DOUBLE NULL DEFAULT 0.0,
 	PRIMARY KEY("ProgramFinancingScheduleId" AUTOINCREMENT)
 );
